tests: move PoW challenge and solution to tests crate to remove cycle
