+++ core/src/ledger/testnet_pow.rs
@@ -489,116 +489,3 @@ mod test {
         assert_eq!(bytes.len(), SOLUTION_VAL_BYTES_LEN);
     }
 }
-
-#[cfg(test)]
-mod test_with_tx_and_vp_env {
-    // IMPORTANT: do not import anything directly from this `crate` here, only
-    // via `namada_tests`. This gets us around the `core -> tests -> core` dep
-    // cycle, which is okay, because `tests` is only a `dev-dependency` of
-    // core and allows us to test the code in the same module as its defined.
-    //
-    // This imports the same code as `super::*` but from a different version of
-    // this crate (one that `namada_tests` depends on). It's re-exported
-    // from `namada_tests` so that we can use it together with
-    // `namada_tests` modules back in here.
-    use namada_tests::namada::core::ledger::storage_api;
-    use namada_tests::namada::core::ledger::testnet_pow::*;
-    use namada_tests::namada::core::types::{address, token};
-    use namada_tests::tx::{self, TestTxEnv};
-    use namada_tests::vp;
-
-    #[test]
-    fn test_challenge_and_solution() -> storage_api::Result<()> {
-        let faucet_address = address::testing::established_address_1();
-        let difficulty = Difficulty::try_new(1).unwrap();
-        let withdrawal_limit = token::Amount::whole(1_000);
-
-        let mut tx_env = TestTxEnv::default();
-
-        // Source address that's using PoW (this would be derived from the tx
-        // wrapper pk)
-        let source = address::testing::established_address_2();
-
-        // Ensure that the addresses exists, so we can use them in a tx
-        tx_env.spawn_accounts([&faucet_address, &source]);
-
-        init_faucet_storage(
-            &mut tx_env.wl_storage,
-            &faucet_address,
-            difficulty,
-            withdrawal_limit,
-        )?;
-        tx_env.commit_genesis();
-
-        let challenge = Challenge::new(
-            &mut tx_env.wl_storage,
-            &faucet_address,
-            source.clone(),
-        )?;
-
-        let solution = challenge.solve();
-
-        // The solution must be valid
-        assert!(solution.verify_solution(source.clone()));
-
-        // Changing the solution to `0` invalidates it
-        {
-            let mut solution = solution.clone();
-            solution.value = 0;
-            // If you're unlucky and this fails, try changing the solution to
-            // a different literal.
-            assert!(!solution.verify_solution(source.clone()));
-        }
-        // Changing the counter invalidates it
-        {
-            let mut solution = solution.clone();
-            solution.params.counter = 10;
-            // If you're unlucky and this fails, try changing the counter to
-            // a different literal.
-            assert!(!solution.verify_solution(source.clone()));
-        }
-
-        // Apply the solution from a tx
-        vp::vp_host_env::init_from_tx(
-            faucet_address.clone(),
-            tx_env,
-            |_addr| {
-                solution
-                    .apply_from_tx(tx::ctx(), &faucet_address, &source)
-                    .unwrap();
-            },
-        );
-
-        // Check that it's valid
-        let is_valid = solution.validate(
-            &vp::ctx().pre(),
-            &faucet_address,
-            source.clone(),
-        )?;
-        assert!(is_valid);
-
-        // Commit the tx
-        let vp_env = vp::vp_host_env::take();
-        tx::tx_host_env::set_from_vp_env(vp_env);
-        tx::tx_host_env::commit_tx_and_block();
-        let tx_env = tx::tx_host_env::take();
-
-        // Re-apply the same solution from a tx
-        vp::vp_host_env::init_from_tx(
-            faucet_address.clone(),
-            tx_env,
-            |_addr| {
-                solution
-                    .apply_from_tx(tx::ctx(), &faucet_address, &source)
-                    .unwrap();
-            },
-        );
-
-        // Check that it's not longer valid
-        let is_valid =
-            solution.validate(&vp::ctx().pre(), &faucet_address, source)?;
-        assert!(!is_valid);
-
-        Ok(())
-    }
-}
