+++ tests/src/storage_api/mod.rs
@@ -1,1 +1,2 @@
 mod collections;
+mod testnet_pow;
